corect signin

+++ my_app/views.py
@@ -8,7 +8,7 @@ def main_page(request):
 def index(request):
     users_list = Users.objects.all().order_by('id')
     context = {'users_list': users_list, 'username': auth.get_user(request).username}
-    if "delete" in request.POST:
+    if "delete" in request.POST and request.user.is_authenticated():
         user_id = request.POST['delete']
         Users.objects.filter(id=user_id).delete()
     return render(request, 'my_app/index.html', context)
@@ -26,9 +26,13 @@ def send_page(request, user):
     return render(request, 'my_app/add.html', {'user': user, 'username': auth.get_user(request).username})
 
 def add(request):
-    user = Users()
-    return send_page(request, user)
+    if request.user.is_authenticated():
+        user = Users()
+        return send_page(request, user)
+    return render(request, 'my_app/permission_denide.html')
     
 def edit(request, id):
-    user = Users.objects.get(id=id)
-    return send_page(request, user)
+    if request.user.is_authenticated():
+        user = Users.objects.get(id=id)
+        return send_page(request, user)
+    return render(request, 'my_app/permission_denide.html')
